split cannon API
